refactor(api): move repository / model to BC campaign-participation
